sharepoint/get_file_from_sh/requirements.txt
SharePlum==0.5.1
requests==2.23.0
python-dotenv==0.13.0
